bot/requirements.txt
python-telegram-bot==22.1
httpx==0.28.1
